persist the order history

+++ app/src/main/java/zoomkart/paykart/models/Item.java
@@ -6,5 +6,5 @@ package zoomkart.paykart.models;
  */
 public class Item {
     public String name;
-    public double price;
+    public float price;
 }
